Create AddTwoNumbers.py
Example 1 
Calculate the sum of two numbers entered by the user..

Example 2
In this program, we asked the user to enter two numbers, and this program displays the sum of the two numbers entered by the user.
We use the built-in input() function to get input data. Since the input() function returns a string, we convert the string to a number using the float() function. Then the numbers are summed up.

Example 3
Alternatively, we can perform this addition in a single statement without using any variables.